+++ test/dummy/app/controllers/recordables_controller.rb
@@ -1,0 +1,5 @@
+class RecordablesController < ApplicationController
+  def index
+    @recordables = [ Workspace, RecordingStudioFolder, RecordingStudioPage, RecordingStudioArchiveBox ]
+  end
+end