+++ test/dummy/config/initializers/recording_studio_api_scalar_docs_moveable_api.rb
@@ -4,5 +4,5 @@ RecordingStudioApi.configure do |config|
   api = config.api(:public)
   api.documentation_enabled = true
   api.documentation_access = :authenticated
-  api.documentation_layout_name = "flat_pack_sidebar"
+  api.documentation_layout_name = "recording_studio/default_layout"
 end